Split out cursor rotation
